+++ dnt.ts
@@ -46,7 +46,11 @@ await build({
     lib: ["ES2021", "DOM"],
   },
   async postBuild() {
-    await Deno.copyFile("logtape/fs.mjs", "npm/esm/fs.js");
+    await Deno.writeTextFile(
+      "npm/esm/fs.js",
+      'import fs from "./fs.cjs";\nexport default fs;\n',
+    );
+    await Deno.copyFile("logtape/fs.cjs", "npm/esm/fs.cjs");
     await Deno.copyFile("logtape/fs.cjs", "npm/script/fs.js");
     await Deno.copyFile("LICENSE", "npm/LICENSE");
     await Deno.copyFile("README.md", "npm/README.md");